+++ libandroid-navigation-ui/src/test/java/com/mapbox/services/android/navigation/ui/v5/NavigationViewOfflineRouterTest.java
@@ -6,7 +6,6 @@ import com.mapbox.services.android.navigation.v5.navigation.OfflineRoute;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
