+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/OfflineRoute.kt
@@ -13,7 +13,7 @@ import okhttp3.HttpUrl
  * could be set in order for an offline navigation session to successfully begin.
  */
 class OfflineRoute private constructor(
-    private val onlineRoute: NavigationRoute,
+    private val onlineRoute: com.mapbox.navigation.route.offboard.NavigationRoute,
     bicycleType: OfflineCriteria.BicycleType?,
     private val cyclingSpeed: Float?,
     private val cyclewayBias: Float?,
@@ -46,7 +46,7 @@ class OfflineRoute private constructor(
          * @return a [Builder] object for creating this object
          */
         @JvmStatic
-        fun builder(onlineRouteBuilder: NavigationRoute.Builder): Builder {
+        fun builder(onlineRouteBuilder: com.mapbox.navigation.route.offboard.NavigationRoute.Builder): Builder {
             return Builder(onlineRouteBuilder)
         }
     }
@@ -118,7 +118,7 @@ class OfflineRoute private constructor(
     private fun obtainRouteFor(response: String): DirectionsRoute? =
         DirectionsResponse.fromJson(response).routes().firstOrNull()
 
-    class Builder internal constructor(private val navigationRouteBuilder: NavigationRoute.Builder) {
+    class Builder internal constructor(private val navigationRouteBuilder: com.mapbox.navigation.route.offboard.NavigationRoute.Builder) {
         private var bicycleType: OfflineCriteria.BicycleType? = null
         private var cyclingSpeed: Float? = null
         private var cyclewayBias: Float? = null
