+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/route/RouteFetcher.kt
@@ -6,7 +6,7 @@ import com.mapbox.api.directions.v5.models.DirectionsResponse
 import com.mapbox.api.directions.v5.models.RouteOptions
 import com.mapbox.geojson.Point
 import com.mapbox.navigation.utils.extensions.ifNonNull
-import com.mapbox.services.android.navigation.v5.navigation.NavigationRoute
+import com.mapbox.navigation.route.offboard.NavigationRoute
 import com.mapbox.services.android.navigation.v5.routeprogress.RouteProgress
 import com.mapbox.services.android.navigation.v5.utils.RouteUtils
 import java.lang.ref.WeakReference
@@ -25,7 +25,7 @@ class RouteFetcher
     context: Context,
     private val accessToken: String,
     private val routeUtils: RouteUtils = RouteUtils(),
-    private var navigationRoute: NavigationRoute? = null
+    private var navigationRoute: com.mapbox.navigation.route.offboard.NavigationRoute? = null
 ) {
 
     companion object {
@@ -100,13 +100,13 @@ class RouteFetcher
     fun buildRequestFrom(
         location: Location,
         routeProgress: RouteProgress
-    ): NavigationRoute.Builder? {
+    ): com.mapbox.navigation.route.offboard.NavigationRoute.Builder? {
         val context = contextWeakReference.get()
             ?: return null
         val origin = Point.fromLngLat(location.longitude, location.latitude)
         val bearing = if (location.hasBearing()) location.bearing.toDouble() else null
         val routeOptions = routeProgress.directionsRoute()?.routeOptions()
-        val navigationRouteBuilder = NavigationRoute.builder(context)
+        val navigationRouteBuilder = com.mapbox.navigation.route.offboard.NavigationRoute.builder(context)
             .accessToken(accessToken)
             .origin(origin, bearing, BEARING_TOLERANCE)
         routeOptions?.let { options ->
@@ -131,7 +131,7 @@ class RouteFetcher
      *
      * @param builder to be executed
      */
-    fun findRouteWith(builder: NavigationRoute.Builder?) {
+    fun findRouteWith(builder: com.mapbox.navigation.route.offboard.NavigationRoute.Builder?) {
         builder?.let { navigationRouteBuilder ->
             navigationRoute = navigationRouteBuilder.build()
             navigationRoute?.getRoute(directionsResponseCallback)
@@ -147,7 +147,7 @@ class RouteFetcher
 
     private fun addDestination(
         remainingWaypoints: MutableList<Point>,
-        builder: NavigationRoute.Builder
+        builder: com.mapbox.navigation.route.offboard.NavigationRoute.Builder
     ) {
         if (remainingWaypoints.isNotEmpty()) {
             builder.destination(retrieveDestinationWaypoint(remainingWaypoints))
@@ -161,7 +161,7 @@ class RouteFetcher
 
     private fun addWaypoints(
         remainingCoordinates: List<Point>,
-        builder: NavigationRoute.Builder
+        builder: com.mapbox.navigation.route.offboard.NavigationRoute.Builder
     ) {
         if (remainingCoordinates.isNotEmpty()) {
             for (coordinate in remainingCoordinates) {
@@ -172,7 +172,7 @@ class RouteFetcher
 
     private fun addWaypointIndices(
         routeProgress: RouteProgress,
-        builder: NavigationRoute.Builder
+        builder: com.mapbox.navigation.route.offboard.NavigationRoute.Builder
     ) {
         val remainingWaypointIndices: IntArray? = routeUtils.calculateRemainingWaypointIndices(routeProgress)
         if (remainingWaypointIndices != null && remainingWaypointIndices.isNotEmpty()) {
@@ -182,7 +182,7 @@ class RouteFetcher
 
     private fun addWaypointNames(
         progress: RouteProgress,
-        builder: NavigationRoute.Builder
+        builder: com.mapbox.navigation.route.offboard.NavigationRoute.Builder
     ) {
         val remainingWaypointNames: Array<String>? = routeUtils.calculateRemainingWaypointNames(progress)
         remainingWaypointNames?.let {
@@ -192,7 +192,7 @@ class RouteFetcher
 
     private fun addApproaches(
         progress: RouteProgress,
-        builder: NavigationRoute.Builder
+        builder: com.mapbox.navigation.route.offboard.NavigationRoute.Builder
     ) {
         val remainingApproaches: Array<String>? = routeUtils.calculateRemainingApproaches(progress)
         remainingApproaches?.let {
