+++ libandroid-navigation/src/test/java/com/mapbox/services/android/navigation/v5/navigation/NavigationRouteTest.kt
@@ -50,7 +50,7 @@ class NavigationRouteTest : BaseTest() {
     @Test
     @Throws(Exception::class)
     fun sanityTest() {
-        val navigationRoute = NavigationRoute.builder(context)
+        val navigationRoute = com.mapbox.navigation.route.offboard.NavigationRoute.builder(context)
             .accessToken(BaseTest.ACCESS_TOKEN)
             .origin(Point.fromLngLat(1.0, 2.0))
             .destination(Point.fromLngLat(1.0, 5.0))
@@ -61,7 +61,7 @@ class NavigationRouteTest : BaseTest() {
     @Test
     @Throws(Exception::class)
     fun changingDefaultValueToCustomWorksProperly() {
-        val navigationRoute = NavigationRoute.builder(context)
+        val navigationRoute = com.mapbox.navigation.route.offboard.NavigationRoute.builder(context)
             .accessToken(BaseTest.ACCESS_TOKEN)
             .origin(Point.fromLngLat(1.0, 2.0))
             .destination(Point.fromLngLat(1.0, 5.0))
@@ -76,7 +76,7 @@ class NavigationRouteTest : BaseTest() {
 
     @Test
     fun addApproachesIncludedInRequest() {
-        val navigationRoute = NavigationRoute.builder(context)
+        val navigationRoute = com.mapbox.navigation.route.offboard.NavigationRoute.builder(context)
             .accessToken(BaseTest.ACCESS_TOKEN)
             .origin(Point.fromLngLat(1.0, 2.0))
             .destination(Point.fromLngLat(1.0, 5.0))
@@ -95,7 +95,7 @@ class NavigationRouteTest : BaseTest() {
 
     @Test
     fun checksWaypointIndicesIncludedInRequest() {
-        val navigationRoute = NavigationRoute.builder(context)
+        val navigationRoute = com.mapbox.navigation.route.offboard.NavigationRoute.builder(context)
             .accessToken(BaseTest.ACCESS_TOKEN)
             .origin(Point.fromLngLat(1.0, 2.0))
             .addWaypoint(Point.fromLngLat(1.0, 3.0))
@@ -112,7 +112,7 @@ class NavigationRouteTest : BaseTest() {
 
     @Test
     fun addWaypointNamesIncludedInRequest() {
-        val navigationRoute = NavigationRoute.builder(context)
+        val navigationRoute = com.mapbox.navigation.route.offboard.NavigationRoute.builder(context)
             .accessToken(BaseTest.ACCESS_TOKEN)
             .origin(Point.fromLngLat(1.0, 2.0))
             .destination(Point.fromLngLat(1.0, 5.0))
@@ -128,7 +128,7 @@ class NavigationRouteTest : BaseTest() {
 
     @Test
     fun addWaypointTargetsIncludedInRequest() {
-        val navigationRoute = NavigationRoute.builder(context)
+        val navigationRoute = com.mapbox.navigation.route.offboard.NavigationRoute.builder(context)
             .accessToken(BaseTest.ACCESS_TOKEN)
             .origin(Point.fromLngLat(1.0, 2.0))
             .destination(Point.fromLngLat(1.0, 5.0))
@@ -143,7 +143,7 @@ class NavigationRouteTest : BaseTest() {
 
     @Test
     fun reverseOriginDestination_bearingsAreFormattedCorrectly() {
-        val navigationRoute = NavigationRoute.builder(context)
+        val navigationRoute = com.mapbox.navigation.route.offboard.NavigationRoute.builder(context)
             .accessToken(BaseTest.ACCESS_TOKEN)
             .destination(Point.fromLngLat(1.0, 5.0), 1.0, 5.0)
             .origin(Point.fromLngLat(1.0, 2.0), 90.0, 90.0)
@@ -156,7 +156,7 @@ class NavigationRouteTest : BaseTest() {
 
     @Test
     fun addWaypointsThenOriginDestination_bearingsAreFormattedCorrectly() {
-        val navigationRoute = NavigationRoute.builder(context)
+        val navigationRoute = com.mapbox.navigation.route.offboard.NavigationRoute.builder(context)
             .accessToken(BaseTest.ACCESS_TOKEN)
             .addWaypoint(Point.fromLngLat(3.0, 4.0), 20.0, 20.0)
             .addWaypoint(Point.fromLngLat(5.0, 6.0), 30.0, 30.0)
@@ -198,7 +198,7 @@ class NavigationRouteTest : BaseTest() {
             )
             .build()
 
-        val navigationRoute = NavigationRoute.builder(context)
+        val navigationRoute = com.mapbox.navigation.route.offboard.NavigationRoute.builder(context)
             .origin(coordinates[0])
             .addWaypoint(coordinates[1])
             .destination(coordinates[2])
@@ -227,7 +227,7 @@ class NavigationRouteTest : BaseTest() {
         val routeCall = mockk<Call<DirectionsResponse>>(relaxed = true)
         every { routeCall.isExecuted } returns false
         every { mapboxDirections.cloneCall() } returns routeCall
-        val navigationRoute = NavigationRoute(mapboxDirections)
+        val navigationRoute = com.mapbox.navigation.route.offboard.NavigationRoute(mapboxDirections)
 
         navigationRoute.cancelCall()
 
@@ -240,7 +240,7 @@ class NavigationRouteTest : BaseTest() {
         val routeCall = mockk<Call<DirectionsResponse>>()
         every { routeCall.isExecuted } returns true
         every { mapboxDirections.cloneCall() } returns routeCall
-        val navigationRoute = NavigationRoute(mapboxDirections)
+        val navigationRoute = com.mapbox.navigation.route.offboard.NavigationRoute(mapboxDirections)
 
         navigationRoute.cancelCall()
 
@@ -250,7 +250,7 @@ class NavigationRouteTest : BaseTest() {
     @Test
     fun builderInterceptor_setsMapboxDirections() {
         val mapboxDirectionsBuilder = mockk<MapboxDirections.Builder>(relaxed = true)
-        val builder = NavigationRoute.Builder(mapboxDirectionsBuilder)
+        val builder = com.mapbox.navigation.route.offboard.NavigationRoute.Builder(mapboxDirectionsBuilder)
         val eventListener = mockk<EventListener>(relaxed = true)
 
         builder.eventListener(eventListener)
@@ -261,7 +261,7 @@ class NavigationRouteTest : BaseTest() {
     @Test
     fun builderEventListener_setsMapboxDirections() {
         val mapboxDirectionsBuilder = mockk<MapboxDirections.Builder>(relaxed = true)
-        val builder = NavigationRoute.Builder(mapboxDirectionsBuilder)
+        val builder = com.mapbox.navigation.route.offboard.NavigationRoute.Builder(mapboxDirectionsBuilder)
         val interceptor = mockk<Interceptor>(relaxed = true)
 
         builder.interceptor(interceptor)
@@ -272,7 +272,7 @@ class NavigationRouteTest : BaseTest() {
     @Test
     fun builderContinueStraight_setsMapboxDirections() {
         val mapboxDirectionsBuilder = mockk<MapboxDirections.Builder>(relaxed = true)
-        val builder = NavigationRoute.Builder(mapboxDirectionsBuilder)
+        val builder = com.mapbox.navigation.route.offboard.NavigationRoute.Builder(mapboxDirectionsBuilder)
         val continueStraight = false
 
         builder.continueStraight(continueStraight)
