+++ libdirections-offboard/build.gradle
@@ -21,6 +21,9 @@ dependencies {
 
     implementation project(':libnavigation-util')
 
+    // Support
+    implementation dependenciesList.supportAppcompatV7
+
     api(project(':libnavigation-base'))
 
     implementation dependenciesList.kotlinStdLib
