+++ libdirections-offboard/src/main/java/com/mapbox/navigation/route/offboard/MapboxOffboardRouter.kt
@@ -14,7 +14,6 @@ class MapboxOffboardRouter : Router {
     }
 
     override fun getRoute(origin: Point, waypoints: List<Point>, listener: Router.RouteListener) {
-
     }
 
     override fun cancel() {
