+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/route/NavigationRouteEventListener.kt
@@ -1,5 +1,6 @@
-package com.mapbox.services.android.navigation.v5.internal.navigation
+package com.mapbox.navigation.base.route
 
+import com.mapbox.navigation.utils.time.ElapsedTime
 import okhttp3.Call
 import okhttp3.EventListener
