+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/route/Router.kt
@@ -7,7 +7,7 @@ interface Router {
     @Deprecated("use [getRoute(Point, List<Point>, RouteListener)] instead")
     fun getRoute(origin: Point, waypoints: List<Point>, callback: (route: Route) -> Unit)
 
-    fun getRoute(origin: Point, waypoints: List<Point>, listener: RouteListener)
+    // fun getRoute(origin: Point, waypoints: List<Point>, listener: RouteListener)
 
     fun cancel()
