+++ libnavigation-util/src/main/java/com/mapbox/navigation/utils/exceptions/NavigationException.kt
